Ignore dummy events in perf2csv and ignore modifiers
When recording with IBS, perf adds a dummy:HG event to the final
data, with no corresponding column in the output of perf report.
This updates the perf2csv script to ignore this event in order not
to use wrong columns for comm, dso, and symbol values. Also, ignore
event modifiers such that "cycles" and "cycles:pp" are both seen as
just "cycles" in the output CSV files.